Rename[]: 추가 Adapter, DBHelper, DTO 패키지
