+++ app/src/main/java/campus/tech/kakao/map/DTO/Meta.kt
@@ -1,7 +1,8 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.DTO
 
 data class Meta(
 	val total_count: Int,
 	val pageable_count: Int,
 	val is_end: Boolean,
-	val same_name: SameName)
+	val same_name: SameName
+)
