+++ app/src/main/java/campus/tech/kakao/map/DTO/PlaceResponse.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.DTO
 
 data class PlaceResponse(
 	val meta: Meta,
