+++ app/src/main/java/campus/tech/kakao/map/DTO/SameName.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.DTO
 
 data class SameName(
 	val region: List<String>,
