+++ app/src/main/java/campus/tech/kakao/map/DTO/SearchWord.kt
@@ -1,3 +1,3 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.DTO
 
 data class SearchWord(val name: String, val address: String, val type: String)
